requirements.txt
sentence-transformers
faiss-cpu
openai
anthropic
python-dotenv